requirements.txt
APScheduler==3.7.0
bcrypt==3.2.0
blis==0.7.4
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
content==0.0.1
cryptography==3.4.6
cycler==0.10.0
cymem==2.0.5
docx==0.2.4
docxpy==0.8.5
en-core-web-sm-abd==0.1
en-core-web-sm-mirror==2.2.5
Flask==1.1.2
google-api-core==1.26.0
google-auth==1.26.1
google-cloud-bigquery==2.10.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.36.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
murmurhash==1.0.5
mysql-connector-python==8.0.23
numpy==1.20.1
packaging==20.9
pandas==1.2.2
paramiko==2.7.2
pathy==0.3.6
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
proto-plus==1.14.2
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
PyMuPDF==1.18.8
PyNaCl==1.4.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.1
schedule==1.1.0
six==1.15.0
smart-open==3.0.0
spacy==2.3.4
spacy-legacy==3.0.1
srsly==1.0.5
storage==0.0.4.3
thinc==7.4.5
tqdm==4.56.2
typer==0.3.2
tzlocal==2.1
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
